Adding form class and merging config codeblocks
2 questions:
* Which other providers are available besides `nucleos_antispam.provider.session`

And what is `antispam_time_max` doing exactly? What happens if I don' t set it: Will there be no max-time, or will there be some (which?) default max-time? Since the max-time is (a) limited by the lifetime of the session (in `php.ini`), and (b) by Symfony's built-in CSRF-protection (which is limited by the session lifetime too), I'm wondering if this is a useful option at all...